generates RBAC Roles for resources that belong to a project.
in practice we are going to generate RBAC for CRDs resources that are part of a project for example "clusters" or "usersshkeieis".
This assures that a user Bob that belongs to a group "readers-projectID" cannot create "cluster" resources.

Notice that this is very different from generating RBAC for a named resources for example a cluster named "my-awesome-cluster".