synchronizes project's resources for seed clusters
since the RBAC Generator application is only run on the master cluster the changes in this pull adjustes the existing code to this requirement.
in practice it means that the app will monitor the project's resources e.g clusters and generate the proper set of RBAC Roles/Bindings for them in
the given seed cluster.