adds migration tool to projects
The purpose of this tool is to find all running clusters and group them by the user they belong to.
Similarly with ssh keys that are being used by the clusters.
For each user the tool creates a default project and starts the migration process by updating the existing resources.