updates authenticated user ID and Name
newly created clusters have its label.user field set to the authenticated user's ID field.
I noticed that there is discrepancy between the authenticated user's ID field and the one we store in the database.
Thus I think the the ID must have been changed and that change was not reflected in the dabase.