removes all leading and trailing white spaces from email addresses
we have duplicate users in your system, the suspects are the email addresses because we use them to look up the users in our system.
The duplicate emails look identical so perhaps at the creation time one of them contained white spaces that were romved before they were stored in the data base.