removes updating users resources
it turns out that the ID taken from the "sub" field of the token is not unique
and changes depending of issuer. For example it will be different when a user logs in
via Google or GitHub even if they map to the same email address. In the new model (UserMgmt)
we don't rely on that value so there is no need to update it every time.

See also: https://app.zenhub.com/workspace/o/kubermatic/kubermatic/issues/2022#issuecomment-424270614