generates a unique name for user resources
previously the name was generated automatically and given the nature of our system that lead to generating multiple object for the same user (email).
To fix the issue we could acquire a distributed lock or generate a unique name. Decided to implement the latter.

the name of the newly created resource will be unique and it is derived from the user's email address (sha256(email)
in the beginning I was considering to hex-encode the email address as it will produce a unique output because the email address in unique.
the only issue I have found with this approach is that the length can get quite long quite fast.
thus decided to use sha256 as it produces fixed output and the hash collisions are very, very, very, very rare.