uses UpdateCloudSpec funtion when updating a cluster
The UpdateCloudSpec is a helper for updating the cloud spec, it rewrites
cloud specific credentials if were not updated.

Note that cloud credentials are removed from responses and they are empty on requests,
for example when a client wants to upgrade a cluster it sends the full cluster spec without the credentials.