Make fluentd crashes visible
So far we've used a fluentd image that contained a supervisor
and on top of that had fluentds own supervising capabiltiy enabled.

That meant that whenever the fluentd worker crashed it would get
restarted, which made the restartCount of the pod stay at zero
even after the process got repeatedly oom-killed by the kernel.