changes kubeconfig file name (#2483)
* changes kubeconfig files name

If the oidc authentication is enabled it is possible to generate kubeconfig
that doesn't contain the admin credentials. The name of the file will be exactly
the same as the one downloaded when clicking on "Download config". To distinguish them
this commit adds a prefix "-admin" to the config with the admin credentials.