prevents inviting a service account to a project
although a service account is a special type of user it should not become a member of a project. This functionality is reserved only for humans :)

service accounts are mainly for programmatic access to the API and will be managed on a dedicated page.
note that this pulls also adds a restriction on an email address a user can have, it cannot start with "serviceaccount-" prefix.