checks projects OwnerRef before deleting a duplicated user (#3343)
Before a user is deleted the tool checks if the user doesn't belong to a project. It first checks user.spec.projects list and if it is empty it also checks UserProjectBindings.
Under normal conditions, the RBAC controller generates UserProjectBindings for project's owners (it examines OnwerReferences).
However a faulty controller might have not created bindings so checking OwnerReferences is the last resort.