adds ci_run_api_e2e scrip for running API E2E test (#3444)
the script is meant to be run in the CI.
It requires `e2e-kind` image that provides deployment and cluster creation scripts.

The script essentially builds a local image that has Kubermatic binaries from the current branch.
Pushes the image to the inner Kubernetes cluster and uses it to install Kubermatic.
Finally, it runs E2E test for the API.