Avoid enqueuing the same cluster twice
`handler.EnqueueRequestForOwner` enqueues the owning object with the
triggering objects namespace, which doesn't work if the owning object
is non-namespaced.

We worked around this by removing the namespace from clusters in the
`Reconcile`. The problem is that we still have two possible queue keys
for the cluster, allowing two workers to concurrently work on the same
cluster, causing a lot of errors.

This PR changes the whole logic to not use the ownerreference but the
namespace of the objects to find the corresponding cluster.