Add credentials-cleanup finalizer after secret creation
Current behaviour is to add the finalizer while deleting the cluster
in the `DeleteEndpoint` method. This works fine if we are deleting
the cluster using the API/dashboard.

But if someone manually issues a `kubectl delete cluster` command, it
does not hit the `DeleteEndpoint` method, due to which the cleanup
finalizer is not added and the secret gets orphaned.

To prevent this scenario, we can directly add the finalizer when we
create the secret. This would also help in cases when `kubectl` is
directly used.