Allow disabling of apiserver endpoints reconciler (#4396)
* Allow disabling of apiserver endpoints reconciler

This change introduces .apiserver.endpointReconcilingDisabled in
componentsOverride of the Cluster CR. When enabled the apiserver
is started with --endpoint-reconciler-type=none which disables
the reconciling of the apiserver endpoint (default: false).
This change also introduces the flag --apiserver-endpoint-reconciling-disabled
for kubermatic-controller-manager which lets the
clusterComponentDefaulter controller default .endpointReconcilingDisabled
to true for all Clusters.

* Add missing inline tag for embedded element.

* Update chart.

* Update chart version.