Add a configurable update window to a cluster
This adds a configurable update window with a start time, (e.g. "02:00", or :"Thu 02:00") and length (e.g. "1h") to a cluster.

It can be used to limit various update operations to a given time frame only.

The first usage is to configure the CoreOS operator node reboot window, see https://github.com/coreos/container-linux-update-operator/blob/master/doc/reboot-windows.md.

In chose to make it a generic setting since we are planning to use this in the future for other things, like automatic cluster upgrades or in the addon controller.

Signed-off-by: Bastian Hofmann <bashofmann@gmail.com>