Better function names for BaseAppLabel and AppClusterLabel (#5150)
* api/pkg: Rename func BaseAppLabel to BaseAppLabels

The function returns a `map[string]string` containing one or more
labels. The singular name `BaseAppLabel` is misleading while the
docstring itself mentions that it returns "labels" and not just a
single "label".

* api/pkg: Rename func AppClusterLabel to AppClusterLabels

The function returns a `map[string]string` containing one or more
labels. The singular name `AppClusterLabel` is misleading while the
docstring itself mentionsthat it returns "labels" and not just a
single "label".