Openstack: include distributed routers in existing router search
When we crate a cluster in an existing subnet, the code looks for
existing routers to which the subnet is connected.

Before this commit we only look for 'network:router_interface'. On some
installations the routers are network:router_interface_distributed.

As a result, Kubermatic user to report no routers found and attempted to
create a new router for the subnet, which is not allowed.