Add Support for Kubernetes 1.18 (#5325)
* fix missing cluster features in addon template data

* refactor user-cluster metrics test to not use port-forwardings anymore

This also includes a fix for the 1.18 compatibility: The kubelet metric
kubelet_runtime_operations_duration_seconds_count replaces
kubelet_runtime_operations_latency_microseconds_count, which was
deprecated since 1.14 and removed in 1.18.

* update Grafana dashboards / Prometheus alerts

This also includes a tiny adjustment for recent yq versions now requiring
--prettyPrint for YAML output when the input was JSON.

* remove code for reconciling pre-1.14 clusters

The intention is to have 1.15 as the minimum supported version, but allow
existing 1.14 clusters to upgrade before we would shred them.

* rebuild test fixtures for 1.15-1.18

* add Kubernetes 1.18 compatibility / bump default version to latest 1.17

The reasoning behind choosing 1.17 over 1.18 as the default is just
feels. Feels that 1.17 has been used in production for longer and
is probably more stable.

* retry ginkgo run when its context was cancelled

Previously, we never really handled the case were the context was
cancelled. One might think that the checks afterwards would keep
us from mis-interpretigin a cancelled timeout as "success":

1. A cancelled ginkgo run does leave behind *some* test reports,
   so the "are there reports at all?" check passes.

2. None of the reports have been marked as failures, because they
   either passed or were still running. So that check also does
   not help us.

With this patch, we will retry correctly.

I also increased the default timeout to a much larger time. I
often encountered timeouts and after increasing the timeout, the
effective times were around 15-17 minutes. So we probably often
kill "good" tests just because the CI nodes were slow.

In the future we might want to export some runtime metrics to
learn more about the individual timings during our e2e tests.

* update/cleanup .prow.yaml

This separates e2e tests into "base" and "extended" tests.

* Base tests run across various k8s versions, but on the same
  cloud provider (AWS) with the same OS (CoreOS).
* Extended tests test the most current supported k8s version
  against various providers/operating systems.

In doing so I removed lots of old (pre-1.15) tests, merged some
and shifted maaany lines around.

The base tests for 1.16-1.18 are always mandatory, the 1.15 is
optional (as it's in a kind-of "suddenly extended live" state).

* improve logging