Fix prometheus alertmanager queries for KubeClientCertificateExpiration (#5737)
In the alert-manager config there is a problem where
`apiserver_client_certificate_expiration_seconds_count{job="apiserver"}
> 0` and `histogram_quantile(0.01, sum by (job, le) (rate(apiserver_client_certificate_expiration_seconds_bucket{job="apiserver"}[5m]))) < 604800`
are joined together by `and`.

However the first portion contains a label `instance` while the second
portion of the query does a `sum(` doesn't include the `instance` label
and therefore prometheus cannot join together those two queries.

This patch adds the instance label to the histogram query in order to
make this alert work.
This can cause an production outage, since you wont be alerted
on a expiring certificate in time.

Signed-off-by: cedi <cedric.kienzler@innovo-cloud.de>

Co-authored-by: cedi <cedric.kienzler@innovo-cloud.de>