Fix bug where cluster in deletion kills controller (#6051)
Sometimes a cluster in deletion can kill the controller possibly
indefinetly, when it tries to add a finalizer to them.
This happens for example when a cluster is in deletion and the
SecretsCleanupFinalizer gets removed when the controller-manager
restarts for some reason.
Then it will always try to add this finalizer again and fail.

Make sure that finalizers only get added when objects are not stuck in
deletion.

Signed-off-by: Phillip Stagnet <p.stagnet@syseleven.de>