Adding a .devcontainer configuration to kubermatic
As per a Discussion I had with @kron4eg I decided to submit a
.devcontainer configuration for kubermatic.

If you never worked with Devcontainers before, here is a good artical by
[Aaron Powell](https://www.aaron-powell.com/about) on why you should care about Devcontainers:
[your open source project needs a devcontainer - here's why](https://www.aaron-powell.com/posts/2021-03-08-your-open-source-project-needs-a-dev-container-heres-why/)

Tl;Dr:
Devcontainers are awesome and they help to define an isolated
development environment which is standardised across all contributors
setups and leverages the beauty of [GitHub Codespaces](https://github.com/features/codespaces) once, it's GA.

Just try it out once, before dismissing it.
Setting up your VScode to use the Devcontainer is easy:
1. Make sure you have the following plugin installed: `ms-vscode-remote.remote-containers`
2. Open VScode in the root-directory of the Repository
3. Click "Reopen in Container" once VScode shows you the popup, telling
you that it detected a devcontainer configuration.

Signed-off-by: Cedric Kienzler <github@cedric-kienzler.de>