handle Openstack project auth variable (#8211)
* preset openstack: handle auth with project/projectID

auth with tenant/TenantID is deprecated inf favor of project/projectID. So in pkg/crd/kubermatic/v1/preset.go we have added the new fields project/projectID.
When we use the preset in code, we frist try to get the Project variables and if there are not defined we fallback to tenantVariables. this logic is implemented
by function pkg/crd/kubermatic/v1/preset:GetProjectOrDefaultToTenant / GetProjectIdOrDefaultToTenantId.

the new preset crd (ie pkg/apis/kubermatic/v1/preset.go) contains only project variables. the pkg/install/crdmigration/clones.go handle the migration of the crd.

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* refactoring OpenstackCredentials to rename Tenant/TenantID to Project/ProjectID

tenant auth id deprecated in openstack in favor of project auth. so we internly store the both auth under project variables.

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* cluster: handle Project/ProjectID for openstack

similar to Preset

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* openstack credential secret: handle project/projectID auth for openstack

the credentials are now store under project variable instead of tenant. to be compatible with existing clusters,
when we read the secret, we first try to read project variables and if there not present, fallback to tenant variables

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* update OpenstackPublicSpec to also return project/ projectID variables

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* update routes /api/v1/providers/openstack/XXX to handle auth project variables.

first try to get auth project varibles in headers and if not defined, fallback to tenant auth variables

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* add missing boilerplate and update generated docs

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* run-api-e2e.sh: create openstack presset with project auth var

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* apply comment from review

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* openstack ceeate machine controller with project auth vars

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>