do not forbid changes to object meta
Previously, the old codebase did this particular validation ONLY in
our REST API. The webhooks never called ValidateUpdateCluster() and
so never checked if metadata was changed.
Now however, the webhooks DO call this function and so it cannot
restrict the changes to metadata.

Funny enough, in handler/common/cluster.go, metadata is not even
changed when updating a cluster. So validating that it doesn't change
is pointless anyway. The API handler does set labels, but those were
then of course magically excluded in the validation.