Update example values and KKPConfig to respect OIDC settings
There were a misleading steps in the example values around the configuration of dex clients and KKPConfiguration.

Value of `issuerClientSecret` belongs to the issuer == used for accessing kubeconfig of user clusters via OIDC authentication.

And thus is actually not needed if users don't use that option.

It was discovered during the investigation of related issue - https://github.com/kubermatic/kubermatic/issues/8723.

Right now the situation was misleading because `issuerClientSecret` is not connected with the usage of `kubermatic` client (that is used for access KKP dahboard).