Machine controller health check extended (#9469)
* machine controller health check extended

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* lint fixed

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* Handling new possible case

If the user cluster client cannot be build because the kubeconfig secret
doesn't exist yet, then simply return StatusDown with no error.

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* error logged

when an error in getting the user cluster client occurs, it is logged
but no returned, because we do not have any way to programmatically get
the reason of the failure.

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* check MC status only if API server is up

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>