ApplicationDefinition: add missing validation check and field for git (#9891)
* applicationDefinition: add missing validation on crd and change .spec.template.source.Git.Ref from string to object

when cloning with go-git api we must specify the type of the reference we clone (branch, tag or commit). so we turn .spec.template.source.Git.Ref to object to have this information

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* applicationDefinition: change crederentials type from GlobalSecretKeySelector to corev1.SecretKeySelector

with PR #9714, the secrets are synced from master to seed only if they have annotation and live in the namespace where kubermatic is installed. Consequently, the namespace can be inferred, so we no longer need GlobalSecretKeySelector.

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* fix lint

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* applicationDefinition add validation on Helm url

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>