Update cluster mutation webhook (#9993)
When cluster is deleted and finalizers are being updated, we have an update trigger that calls default appliers.
This is not needed in a case when some clusters in defaulting requires to have an access to cloud provider over a secret that might be not available anymore (KubeVirt use case).

Signed-off-by: Sankalp Rangare <sankalprangare786@gmail.com>