application: donwload sources in an emptyDir
the size of the empty dir can be set through Cluster.spec.applicationSettings.CacheSize. The default size is 300Mi.

To choose this default value, we looks a the size of 3 repositories:
* https://github.com/bitnami/charts: set of ~100 charts provided by bitnami
* https://github.com/prometheus-community/helm-charts: contains the different charts for deploying prometheus stack
* https://github.com/kubermatic/kubermatic

These repositories has been choosen because, the first 2 serve only charts and the last one serve charts and code (with history).

repository              size with clone commit          size with clone tag
bitnami                 199m                            38m
prometheus              87m                             6,7m
kubermatic              191m                            50m

So with a 300Mi cache should be more than enough to donwload source from a big repo and handle dependencies

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>