Applications fix panic in HelmClient.DonwloadChart when version is empty or  non semver (#10662)
* rename pkg/applications/providers/util/util.go:AuthFromCredentials  to HelmAuthFromCredentials

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* HelmClient: fix panic when Downloading chart from oci repository with a empty or non semVersion

problem was that when version is not a semVersion. Helm will list tag using ChartDownloader.RegistryClient and not the one set in option. As we did'nt set it we got a NPE.

according to vendor/helm.sh/helm/v3/pkg/downloader/chart_downloader.go.ResolveChartVersion(reg, version string) version must be a semVersion or an empty string. If it's an empty string then last char version is returned.
So we have adapt tests accordingly.

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>