Anexia: allow extended disk configuration in AnexiaNodeSpec
This extends apiv1.AnexiaNodeSpec to allow extended disk configuration
introduced in machine-controller with pull request 1402
(https://github.com/kubermatic/machine-controller/pull/1402).

Without this, Kubermatic API returns a 500 when trying to do any
MachineDeployment operation on a cluster which has a MachineDeployment
with the new disks attribute (and thus without the old diskSize
attribute).

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>