application: gitSource, avoid confusiing message if git remote is using file protocol (#10858)
we use go-git libray which is describe as 'a highly extensible git implementation library written in pure Go' but if you read carrefully https://github.com/go-git/go-git/blob/master/COMPATIBILITY.md, you can see library fallback to the git binary.

if the user make a mistake in the git remote url. it can be intergreted as remote with file protocol and raise the confusing error: "'git' executable not found in $PATH".  To avoid that we cath the error and say the file protocol is not supported.

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>