conformance-tester, scenario Anexia: fix setting cloud-init in OS Spec
Setting the OperatingSystemSpec on the NodeSpec given into
machine.GetAnexiaProviderConfig() does not actually do anything with the
OS spec, instead, when creating the MachineDeployment, another one is
created - completely ignoring us having set the ProvisioningUtility to
CloudInit.

This fixes that problem by copying the relevant code from the
baseScenario.createMachineDeployment method and modifying the OS spec
given to the createMachineDeployment() function.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>