Anexia: allow extended disk configuration in AnexiaNodeSpec (#10816)
* Anexia: allow extended disk configuration in AnexiaNodeSpec

This extends apiv1.AnexiaNodeSpec to allow extended disk configuration
introduced in machine-controller with pull request 1402
(https://github.com/kubermatic/machine-controller/pull/1402).

Without this, Kubermatic API returns a 500 when trying to do any
MachineDeployment operation on a cluster which has a MachineDeployment
with the new disks attribute (and thus without the old diskSize
attribute).

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: disable OSM in E2E tests

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: override default usage of ignition in e2e tests

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* conformance-tester, scenario Anexia: fix setting cloud-init in OS Spec

Setting the OperatingSystemSpec on the NodeSpec given into
machine.GetAnexiaProviderConfig() does not actually do anything with the
OS spec, instead, when creating the MachineDeployment, another one is
created - completely ignoring us having set the ProvisioningUtility to
CloudInit.

This fixes that problem by copying the relevant code from the
baseScenario.createMachineDeployment method and modifying the OS spec
given to the createMachineDeployment() function.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* conformance-tester/baseScenario: correctly encode OperatingSystemSpec

OperatingSystemSpec directly contains the config attributes for the
given operating system, but was wrong encoded as
{"$operatingSystem": {$OperatingSystemSpec...}}

This fixes that issue by adding another switch-case over all known
operating systems. Should probably be done in a better way, but need
this done already..

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>