remove error decoding
smithy.APIError implementations all seem to be doing the same, if not
better string marshalling that we would do.

Also using HTTP 500 (internal server error) for every possible fault
is not ideal (ValidateCredentials() should not error out with an internal
server error on bad credentials), and in general the code in the
provider package should not return HTTP-related things, as controllers
also use it.

I removed the DecodeError() entirely, as it seems to serve no purpose
anymore with SDK v2. :)