applications: add template creadentials to applicationTemplate
In some case credentials may be required to "template" the application Source.

Example:
if the template method is Helm and Source is GIT, in order to install the chart, we must download the chart dependencies defined in Chart.yaml. These dependencies may point to a repository that requires authentication.

of course if the source is helm, we don't  need "template credentials" as dependencies are packaged in the chart archive

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>