Applications: download helm dependencies with credentials (#11452)
* move pkg/applications/helmclient/client_integration_test.go::copyDir func to pkg/applications/test/helm.go::CopyDir

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* refactor pkg/applications/providers/template/helm_integration_test.go::installorUpgradeTest() to take chart location as parmeter instead of global variable

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* applications: add template creadentials to applicationTemplate

In some case credentials may be required to "template" the application Source.

Example:
if the template method is Helm and Source is GIT, in order to install the chart, we must download the chart dependencies defined in Chart.yaml. These dependencies may point to a repository that requires authentication.

of course if the source is helm, we don't  need "template credentials" as dependencies are packaged in the chart archive

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* applications: update crd documentation  example

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* fix typo

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* apply review comments

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>