make reconciling for clusters with LoadBalancer expose strategy clearer
The previous implementation of the TemplateData struct implicitly required Services to
exist and hid the fact it required cluster access by using the evil, embedded, legacy
context/kubeclient. This alone breaks some linters that forbid contexts to be embedded
into structs, but sadly we have amassed some code already that uses this pattern. New code
however should strive to avoid this.

This commit cleans up a recent addition by moving the necessary logic into the controller
and, more importantly, leaving a comment that due to the asynchronous nature of CCMs, the
front-loadbalancer's Service won't have a status immediately after they were created. This
means re-reconciling is necessary in any case to catch updates to the Services.

So moving the code that determines the alt names *before* the Services have been created
(in order to build a single data object) is not making the number of reconciliations worse,
but makes it more clear that they happen by having the related code right in the controller.
Another option would have been to have a simple Setter() function on the data to inject the
altnames even after the data struct has been built already. But again, multiple rounds of
reconciliations are necessary anyway, so we can keep the code in one place.