fix missing secret
Since #13703 (merged shortly after 2.26 branched off) the deployment reconcilers don't call
GetPodTemplateLabels(), but instead let the VolumeRevisionLabelModifier take care of that. So in
the main branch there is no need anymore to create dummy secrets, so #14063 never had to create one.

In the 2.26 branch however, those still need to be maintained.