Stop adding inappropriate finalizer in openstack provider
Previously we incorrectly added kubermatic.k8c.io/cleanup-openstack-router-subnet-link-v2 finalizer
on the event when we created the subnet. And then again when we link a subnet to the router. This
is a problem since we have a logic that can skip linking to the router, but the finalizer is mistakenly
added anyway.

Signed-off-by: Artiom Diomin <artiom@kubermatic.com>